+++ test/ILLink.RoslynAnalyzer.Tests/StaticInterfaceMethods.cs
@@ -1,4 +1,7 @@
-﻿using System;
+﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
 using System.Threading.Tasks;
 using Xunit;
 
@@ -6,56 +9,54 @@ namespace ILLink.RoslynAnalyzer.Tests.Inheritance.Interfaces
 {
 	public sealed partial class StaticInterfaceMethodsTests : LinkerTestBase
 	{
-
 		protected override string TestSuiteName => "Inheritance.Interfaces.StaticInterfaceMethods";
 
 		[Fact]
 		public Task StaticAbstractInterfaceMethods ()
 		{
-			return RunTest (allowMissingWarnings: true);
+			return RunTest (allowMissingWarnings: false);
 		}
 
 		[Fact]
 		public Task StaticAbstractInterfaceMethodsLibrary ()
 		{
-			return RunTest (allowMissingWarnings: true);
+			return RunTest (allowMissingWarnings: false);
 		}
 
 		[Fact]
 		public Task StaticInterfaceMethodsInPreservedScope ()
 		{
-			return RunTest (allowMissingWarnings: true);
+			return RunTest (allowMissingWarnings: false);
 		}
 
 		[Fact]
 		public Task StaticVirtualInterfaceMethodsInPreservedScope ()
 		{
-			return RunTest (allowMissingWarnings: true);
+			return RunTest (allowMissingWarnings: false);
 		}
 
 		[Fact]
 		public Task StaticVirtualInterfaceMethodsInPreservedScopeLibrary ()
 		{
-			return RunTest (allowMissingWarnings: true);
+			return RunTest (allowMissingWarnings: false);
 		}
 
 		[Fact]
 		public Task StaticVirtualInterfaceMethodsLibrary ()
 		{
-			return RunTest (allowMissingWarnings: true);
+			return RunTest (allowMissingWarnings: false);
 		}
 
 		[Fact]
 		public Task UnusedInterfacesInPreservedScope ()
 		{
-			return RunTest (allowMissingWarnings: true);
+			return RunTest (allowMissingWarnings: false);
 		}
 
 		[Fact]
 		public Task UnusedStaticInterfaceMethods ()
 		{
-			return RunTest (allowMissingWarnings: true);
+			return RunTest (allowMissingWarnings: false);
 		}
-
 	}
 }
