--- test/Mono.Linker.Tests.Cases/Inheritance.Interfaces/StaticInterfaceMethods/StaticVirtualInterfaceMethods.cs
@@ -1,65 +1,0 @@
-﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Mono.Linker.Tests.Cases.Expectations.Assertions;
-using Mono.Linker.Tests.Cases.Expectations.Helpers;
-using Mono.Linker.Tests.Cases.Expectations.Metadata;
-using Mono.Linker.Tests.Cases.Inheritance.Interfaces.StaticInterfaceMethods.Dependencies;
-
-namespace Mono.Linker.Tests.Cases.Inheritance.Interfaces.StaticInterfaceMethods
-{
-	[SetupCompileBefore ("library.dll", new[] { "Dependencies/Library.cs" })]
-	[SetupLinkerAction ("skip", "library")]
-	[SetupLinkerArgument ("-a", "test.exe")]
-	public static class StaticVirtualInterfaceMethods
-	{
-		[Kept]
-		public static void Main ()
-		{
-			var x = typeof (IfaceMethodInPreserveScope.ExplcitImplementations);
-			x = typeof (IfaceMethodInPreserveScope);
-			x = typeof (AbstractStaticInPreserveScope.ExplcitImplementations);
-			x = typeof (AbstractStaticInPreserveScope);
-		}
-
-		[Kept]
-		public static class IfaceMethodInPreserveScope
-		{
-			[Kept]
-			[KeptInterface (typeof (IStaticVirtualMethods))]
-			public class ExplcitImplementations : IStaticVirtualMethods
-			{
-				[Kept]
-				static int IStaticVirtualMethods.Property { [Kept][KeptOverride (typeof (IStaticVirtualMethods))] get => 1; [Kept][KeptOverride (typeof (IStaticVirtualMethods))] set => _ = value; }
-				[Kept]
-				[KeptOverride (typeof (IStaticVirtualMethods))]
-				static int IStaticVirtualMethods.Method () => 1;
-				int IStaticVirtualMethods.InstanceMethod () => 0;
-			}
-		}
-
-		[Kept]
-		public static class AbstractStaticInPreserveScope
-		{
-			[Kept]
-			[KeptInterface (typeof (IStaticAbstractMethods))]
-			public class ExplcitImplementations : IStaticAbstractMethods
-			{
-				[Kept]
-				static int IStaticAbstractMethods.Property { [Kept][KeptOverride (typeof (IStaticAbstractMethods))] get => 1; [Kept][KeptOverride (typeof (IStaticAbstractMethods))] set => _ = value; }
-				[Kept]
-				[KeptOverride (typeof (IStaticAbstractMethods))]
-				static int IStaticAbstractMethods.Method () => 1;
-				[Kept]
-				[KeptOverride (typeof (IStaticAbstractMethods))]
-				int IStaticAbstractMethods.InstanceMethod () => 0;
-			}
-		}
-	}
-}
-
