+++ test/Mono.Linker.Tests.Cases/Inheritance.Interfaces/StaticInterfaceMethods/StaticVirtualInterfaceMethodsInPreservedScopeLibrary.cs
@@ -15,77 +15,73 @@ namespace Mono.Linker.Tests.Cases.Inheritance.Interfaces.StaticInterfaceMethods
 {
 	[SetupCompileBefore ("library.dll", new[] { "Dependencies/Library.cs" })]
 	[SetupLinkerAction ("skip", "library")]
-	[SetupLinkerArgument ("-a", "test.exe")]
-	public static class StaticVirtualInterfaceMethodsInPreserveScope
+	[SetupLinkerArgument ("-a", "test.exe", "library")]
+	public static class StaticVirtualInterfaceMethodsInPreservedScopeLibrary
 	{
 		[Kept]
 		public static void Main ()
 		{
-			NotRelevantToVariantCasting.Keep ();
-			var t = typeof (RelevantToVariantCasting);
-			MarkInterfaceMethods<UsedAsTypeArgument> ();
-			var x = new InstantiatedClass ();
 		}
 
-		[Kept]
-		static void MarkInterfaceMethods<T> () where T : IStaticInterfaceWithDefaultImpls
-		{
-			T.Property = T.Property + 1;
-			T.Method ();
-			CallInstanceMethod (null);
-
-			[Kept]
-			void CallInstanceMethod (IStaticInterfaceWithDefaultImpls x)
-			{
-				x.InstanceMethod ();
-			}
-		}
-
-		[Kept]
-		[KeptInterface (typeof (IStaticInterfaceWithDefaultImpls))]
-		public class RelevantToVariantCasting : IStaticInterfaceWithDefaultImpls
-		{
-			[Kept]
-			static int IStaticInterfaceWithDefaultImpls.Property { [Kept][KeptOverride (typeof (IStaticInterfaceWithDefaultImpls))] get => 1; [Kept][KeptOverride (typeof (IStaticInterfaceWithDefaultImpls))] set => _ = value; }
-			[Kept]
-			[KeptOverride (typeof (IStaticInterfaceWithDefaultImpls))]
-			static int IStaticInterfaceWithDefaultImpls.Method () => 1;
-			int IStaticInterfaceWithDefaultImpls.InstanceMethod () => 0;
-		}
-
-		[Kept]
-		[KeptInterface (typeof (IStaticInterfaceWithDefaultImpls))]
-		public class UsedAsTypeArgument : IStaticInterfaceWithDefaultImpls
-		{
-			[Kept]
-			static int IStaticInterfaceWithDefaultImpls.Property { [Kept][KeptOverride (typeof (IStaticInterfaceWithDefaultImpls))] get => 1; [Kept][KeptOverride (typeof (IStaticInterfaceWithDefaultImpls))] set => _ = value; }
-			[Kept]
-			[KeptOverride (typeof (IStaticInterfaceWithDefaultImpls))]
-			static int IStaticInterfaceWithDefaultImpls.Method () => 1;
-			int IStaticInterfaceWithDefaultImpls.InstanceMethod () => 0;
-		}
-
-		[Kept]
-		public class NotRelevantToVariantCasting : IStaticInterfaceWithDefaultImpls
-		{
-			[Kept]
-			public static void Keep () { }
-			static int IStaticInterfaceWithDefaultImpls.Property { get => 1; set => _ = value; }
-			static int IStaticInterfaceWithDefaultImpls.Method () => 1;
-			int IStaticInterfaceWithDefaultImpls.InstanceMethod () => 0;
-		}
 		[Kept]
 		[KeptMember (".ctor()")]
 		[KeptInterface (typeof (IStaticInterfaceWithDefaultImpls))]
-		public class InstantiatedClass : IStaticInterfaceWithDefaultImpls
+		public class ImplementVirtualIface : IStaticInterfaceWithDefaultImpls
 		{
-			//[Kept] //Should be able to remove if not relevant to variant casting
-			static int IStaticInterfaceWithDefaultImpls.Property { /*[Kept][KeptOverride (typeof (IStaticInterfaceWithDefaultImpls))]*/ get => 1; /*[Kept][KeptOverride (typeof (IStaticInterfaceWithDefaultImpls))]*/ set => _ = value; }
-			//[Kept] //Should be able to remove if not relevant to variant casting
-			//[KeptOverride (typeof (IStaticInterfaceWithDefaultImpls))]
+			[Kept]
+			static int IStaticInterfaceWithDefaultImpls.Property { [Kept][KeptOverride (typeof (IStaticInterfaceWithDefaultImpls))] get => 1; [Kept][KeptOverride (typeof (IStaticInterfaceWithDefaultImpls))] set => _ = value; }
+			[Kept]
+			[KeptOverride (typeof (IStaticInterfaceWithDefaultImpls))]
 			static int IStaticInterfaceWithDefaultImpls.Method () => 1;
 			[Kept]
+			[KeptOverride (typeof (IStaticInterfaceWithDefaultImpls))]
 			int IStaticInterfaceWithDefaultImpls.InstanceMethod () => 0;
+		}
+
+		[Kept]
+		[KeptInterface (typeof (IStaticInterfaceWithDefaultImpls))]
+		public class ImplementVirtualIfaceProtectedCtor : IStaticInterfaceWithDefaultImpls
+		{
+			[Kept]
+			protected ImplementVirtualIfaceProtectedCtor () { }
+			[Kept]
+			static int IStaticInterfaceWithDefaultImpls.Property { [Kept][KeptOverride (typeof (IStaticInterfaceWithDefaultImpls))] get => 1; [Kept][KeptOverride (typeof (IStaticInterfaceWithDefaultImpls))] set => _ = value; }
+			[Kept]
+			[KeptOverride (typeof (IStaticInterfaceWithDefaultImpls))]
+			static int IStaticInterfaceWithDefaultImpls.Method () => 1;
+			[Kept]
+			[KeptOverride (typeof (IStaticInterfaceWithDefaultImpls))]
+			int IStaticInterfaceWithDefaultImpls.InstanceMethod () => 0;
+		}
+
+		[Kept]
+		[KeptInterface (typeof (IStaticInterfaceWithDefaultImpls))]
+		public class ImplementVirtualIfaceUninstantiated : IStaticInterfaceWithDefaultImpls
+		{
+			private ImplementVirtualIfaceUninstantiated () { }
+			[Kept]
+			static int IStaticInterfaceWithDefaultImpls.Property { [Kept][KeptOverride (typeof (IStaticInterfaceWithDefaultImpls))] get => 1; [Kept][KeptOverride (typeof (IStaticInterfaceWithDefaultImpls))] set => _ = value; }
+			[Kept]
+			[KeptOverride (typeof (IStaticInterfaceWithDefaultImpls))]
+			static int IStaticInterfaceWithDefaultImpls.Method () => 1;
+			// Type has private ctor, so instance methods can be removed. Since there's a default impl, we can remove this interface method
+			[Kept]
+			[KeptOverride (typeof (IStaticInterfaceWithDefaultImpls))]
+			int IStaticInterfaceWithDefaultImpls.InstanceMethod () => 0;
+		}
+
+		[Kept]
+		[KeptInterface (typeof (IStaticInterfaceWithDefaultImpls))]
+		public class ImplicitImplementVirtualIfaceUninstantiated : IStaticInterfaceWithDefaultImpls
+		{
+			private ImplicitImplementVirtualIfaceUninstantiated () { }
+			[Kept]
+			public static int Property { [Kept][KeptOverride (typeof (IStaticInterfaceWithDefaultImpls))] get => 1; [Kept][KeptOverride (typeof (IStaticInterfaceWithDefaultImpls))] set => _ = value; }
+			[Kept]
+			[KeptOverride (typeof (IStaticInterfaceWithDefaultImpls))]
+			public static int Method () => 1;
+			[Kept]
+			public int InstanceMethod () => 0;
 		}
 	}
 }
