+++ test/Mono.Linker.Tests.Cases/Inheritance.Interfaces/StaticInterfaceMethods/VarianceBasic.cs
@@ -6,18 +6,6 @@ using Mono.Linker.Tests.Cases.Expectations.Metadata;
 
 namespace Mono.Linker.Tests.Cases.Inheritance.Interfaces.StaticInterfaceMethods
 {
-	//[SetupLinkerArgument ("--skip-unresolved", "true")]
-	//[TestCaseRequirements (TestRunCharacteristics.SupportsStaticInterfaceMethods, "Requires a framework that supports static interface methods")]
-	//[Define ("IL_ASSEMBLY_AVAILABLE")]
-	//[SetupCompileBefore ("library.dll", new[] { "Dependencies/VarianceBasic.il" })]
-
-	//[KeptTypeInAssembly ("library.dll", "InterfaceScenario1`1")]
-	//[KeptTypeInAssembly ("library.dll", "InterfaceScenario2`1")]
-	//[KeptMemberInAssembly ("library.dll", "BaseScenario1", "Method()")]
-	//[KeptMemberInAssembly ("library.dll", "BaseScenario2", "Method()", "Method2()", "Method_Obj()", "Method2_Obj()")]
-	//[KeptTypeInAssembly ("library.dll", "DerivedScenario1")]
-	//[KeptMemberInAssembly ("library.dll", "DerivedScenario2", "Method()", "Method2()")]
-
 	class VarianceBasic
 	{
 		[Kept]
